# shared/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Concierge-Career API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/profile": {
      "post": {
        "summary": "Create Profile",
        "operationId": "create_profile_api_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile/{user_id}": {
      "get": {
        "summary": "Get Profile",
        "operationId": "get_profile_api_profile__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Profile",
        "operationId": "update_profile_api_profile__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile/{user_id}/skills": {
      "post": {
        "summary": "Add Skill",
        "operationId": "add_skill_api_profile__user_id__skills_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/reset/{user_id}": {
      "post": {
        "summary": "Reset",
        "operationId": "reset_api_demo_reset__user_id__post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resume/analyze": {
      "post": {
        "summary": "Resume",
        "operationId": "resume_api_resume_analyze_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_resume_api_resume_analyze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/analysis/recalculate": {
      "post": {
        "summary": "Analyze",
        "operationId": "analyze_api_analysis_recalculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/career": {
      "post": {
        "summary": "Analyze",
        "operationId": "analyze_api_analysis_career_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "summary": "Jobs",
        "operationId": "jobs_api_jobs_get",
        "parameters": [
          {
            "name": "career_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Career Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "summary": "Job",
        "operationId": "job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/match": {
      "get": {
        "summary": "Job Match",
        "operationId": "job_match_api_jobs__job_id__match_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobMatchOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}/skill-gaps": {
      "get": {
        "summary": "Gaps",
        "operationId": "gaps_api_users__user_id__skill_gaps_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "career_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Career Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GapOutput"
                  },
                  "title": "Response Gaps Api Users  User Id  Skill Gaps Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}/career-readiness": {
      "get": {
        "summary": "Readiness",
        "operationId": "readiness_api_users__user_id__career_readiness_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "career_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Career Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}/alternative-careers": {
      "get": {
        "summary": "Alternatives",
        "operationId": "alternatives_api_users__user_id__alternative_careers_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "career_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Career Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlternativeOutput"
                  },
                  "title": "Response Alternatives Api Users  User Id  Alternative Careers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/learning-recommendations": {
      "get": {
        "summary": "Learning",
        "operationId": "learning_api_learning_recommendations_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "career_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Career Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LearningOutput"
                  },
                  "title": "Response Learning Api Learning Recommendations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/progress/complete": {
      "post": {
        "summary": "Complete",
        "operationId": "complete_api_progress_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}/saved-jobs": {
      "get": {
        "summary": "Saved",
        "operationId": "saved_api_users__user_id__saved_jobs_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}/saved-jobs/{job_id}": {
      "put": {
        "summary": "Save Job",
        "operationId": "save_job_api_users__user_id__saved_jobs__job_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Unsave Job",
        "operationId": "unsave_job_api_users__user_id__saved_jobs__job_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/concierge/chat": {
      "post": {
        "summary": "Chat",
        "operationId": "chat_api_concierge_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProfileInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "university": {
            "type": "string",
            "maxLength": 200,
            "title": "University",
            "default": ""
          },
          "field": {
            "type": "string",
            "maxLength": 200,
            "title": "Field",
            "default": ""
          },
          "degree": {
            "type": "string",
            "maxLength": 100,
            "title": "Degree",
            "default": "Bachelor"
          },
          "graduation_year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 1950.0,
            "title": "Graduation Year",
            "default": 2027
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillInput"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Skills"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationInput"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Education"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/ExperienceInput"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Experiences"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectInput"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Projects"
          },
          "certifications": {
            "items": {
              "$ref": "#/components/schemas/CertificationInput"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Certifications"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProfileInput"
      },
      "ProfileOutput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "university": {
            "type": "string",
            "maxLength": 200,
            "title": "University",
            "default": ""
          },
          "field": {
            "type": "string",
            "maxLength": 200,
            "title": "Field",
            "default": ""
          },
          "degree": {
            "type": "string",
            "maxLength": 100,
            "title": "Degree",
            "default": "Bachelor"
          },
          "graduation_year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 1950.0,
            "title": "Graduation Year",
            "default": 2027
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillInput"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Skills"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationInput"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Education"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/ExperienceInput"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Experiences"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectInput"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Projects"
          },
          "certifications": {
            "items": {
              "$ref": "#/components/schemas/CertificationInput"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Certifications"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "revision"
        ],
        "title": "ProfileOutput"
      },
...
    }
  }
}